+++ frontend/components/Layouts/LandingPageLayout.js
@@ -1,0 +1,11 @@
+import Nav from '../Site/Nav'
+
+const LandingPageLayout = props => (
+  <div>
+    <Nav />
+
+    {props.children}
+  </div>
+)
+
+export default LandingPageLayout
