+++ frontend/components/Site/HomePage/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './HomePage'
