--- frontend/components/Site/Testimonials/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Testimonials'
